% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chunk.r
\name{add_chunk}
\alias{add_chunk}
\title{Add a chunk to the disk.frame}
\usage{
add_chunk(df, chunk, chunk_id = NULL, full.names = FALSE)
}
\arguments{
\item{df}{the disk.frame to add a chunk to}

\item{chunk}{a data.frame to be added as a chunk}

\item{chunk_id}{a numeric number indicating the id of the chunk. If NULL it will be set to the largest chunk_id + 1}

\item{full.names}{whether the chunk_id name match should be to the full file path not just the file name}
}
\description{
Add a chunk to the disk.frame
}
