% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.r
\name{as.data.table.disk.frame}
\alias{as.data.table.disk.frame}
\title{Convert disk.frame to data.table by collecting all chunks}
\usage{
\method{as.data.table}{disk.frame}(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{a disk.frame}

\item{keep.rownames}{passed to as.data.table}

\item{...}{passed to as.data.table}
}
\description{
Convert disk.frame to data.table by collecting all chunks
}
\examples{
library(data.table)
cars.df = as.disk.frame(cars)
as.data.table(cars.df)

# clean up
delete(cars.df)
}
