% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.disk.frame.r, R/hard_group_by.r
\name{as.disk.frame}
\alias{as.disk.frame}
\title{Make a data.frame into a disk.frame}
\usage{
as.disk.frame(df, outdir, nchunks = recommend_nchunks(df),
  overwrite = F, compress = 50, ...)

as.disk.frame(df, outdir, nchunks = recommend_nchunks(df),
  overwrite = F, compress = 50, ...)
}
\arguments{
\item{df}{a disk.frame}

\item{outdir}{the output directory}

\item{nchunks}{number of chunks}

\item{overwrite}{if TRUE the outdir will be overwritten, if FALSE it will throw an error if the directory is not empty}

\item{compress}{the compression level 0-100; 100 is highest}

\item{...}{passed to output_disk.frame}

\item{df}{a disk.frame}

\item{outdir}{the output directory}

\item{nchunks}{number of chunks}

\item{overwrite}{if TRUE the outdir will be overwritten, if FALSE it will throw an error if the directory is not empty}

\item{compress}{the compression level 0-100; 100 is highest}

\item{...}{passed to output_disk.frame}
}
\description{
Make a data.frame into a disk.frame

Make a data.frame into a disk.frame
}
