% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.r
\name{auc}
\alias{auc}
\title{compute Area under the Curve (AUC)}
\usage{
auc(target, score)
}
\arguments{
\item{target}{the target}

\item{score}{the score, the higher the score the less targets}
}
\description{
compute Area under the Curve (AUC)
}
