% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloom-filter.r
\name{bf_likely_in_chunks}
\alias{bf_likely_in_chunks}
\title{Return the chunks that values are in certain chunks}
\usage{
bf_likely_in_chunks(df, cols, values)
}
\description{
Return the chunks that values are in certain chunks
}
