% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{chunk_lapply}
\alias{chunk_lapply}
\title{The first chunk of the disk.frame
firstchunk <- function(df) {
  eval(parse(text = sprintf("attr(%s,'chunk.seq') <- 1",as.character(substitute(df)) )),envir = parent.frame())
  attr(df, "chunk.seq") <- 1
  fst::read.fst(attr(df,"path"), from = 1, to = 1000000)
}}
\usage{
chunk_lapply(df, fn, ..., outdir = NULL, chunks = 16, compress = 100)
}
\description{
#' Gives the next chunk of the disk.frame
nextchunk <- function(df) {
  # df is a disk.frame chunk
  fromi <- (attr(df,"chunk.seq")+1)*1000000+1
  toi <- fromi + 1000000 -1
  
  dfstr <- as.character(substitute(df))
  eval(parse(text = sprintf("attr(%s,'chunk.seq') <- attr(%s,'chunk.seq') + 1 ", dfstr, dfstr)), envir = parent.frame())
  fst::read.fst(attr(df,"path"), from = fromi, to = toi)
}
}
\details{
start <- function(df) {
  
}
do to all chunks
}
