% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{chunk_reader_fn}
\alias{chunk_reader_fn}
\title{creates function that will read a file in successive parts using readLines.}
\usage{
chunk_reader_fn(infile, nchunks = 1, num_header_rows = 0)
}
\description{
creates function that will read a file in successive parts using readLines.
}
\keyword{internal}
