% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{chunk_reader_fn_readr}
\alias{chunk_reader_fn_readr}
\title{creates callback function that can be used with readr::read_lines_chunked}
\usage{
chunk_reader_fn_readr(reader_fn, num_header_rows = 0, ...)
}
\description{
creates callback function that can be used with readr::read_lines_chunked
}
\keyword{internal}
