% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{chunks_lapply}
\alias{chunks_lapply}
\title{Access an element of the disk.frame 
`$.disk.frame`<- function(x,y) {
  xx <- x
  class(xx) <- "list"
  fst::read.fst(.Primitive("$")(xx, "path"), columns = substitute(y))
}}
\usage{
chunks_lapply(df, fn, ..., outdir = NULL, chunks = 16, parallel = F)
}
\description{
#' The [ method of disk.frame
`[.disk.frame` <- function(df, chunk_id = -1, ...) {
  # b <- data.table::fread(attr(df,"file"),na.strings=".")
  # chunk_id <- -1 means all the chunks
  env <- attr(df,"env")
  env$data[...]
}
}
