% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.r
\name{collect.disk.frame}
\alias{collect.disk.frame}
\alias{collect_list}
\title{Bring the disk.frame into R as data.table/data.frame}
\usage{
\method{collect}{disk.frame}(x, ..., parallel = !is.null(attr(x,
  "lazyfn")))

collect_list(x, simplify = F, parallel = !is.null(attr(x, "lazyfn")))
}
\arguments{
\item{x}{a disk.frame}

\item{...}{not used}

\item{parallel}{if TRUE the collection is performed in parallel. By default
if there are delayed/lazy steps then it will be parallel, otherwise it will
not be in parallel. This is because parallel requires transferring data
from background R session to the current R session and if there is no
computation then it's better to avoid transferring data between session,
hence parallel = FALSE is a better choice}

\item{simplify}{Should the result be simplified to array}
}
\description{
Bring the disk.frame into R as data.table/data.frame

Bring the disk.frame into R as list
}
