% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{compat_as_lazy}
\alias{compat_as_lazy}
\title{Internal method replicating dplyr's compat_as_lazy_dots as it's not exported}
\usage{
compat_as_lazy(quo)
}
\arguments{
\item{quo}{a quote expression}
}
\description{
Internal method replicating dplyr's compat_as_lazy_dots as it's not exported
}
