% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{compat_as_lazy_dots}
\alias{compat_as_lazy_dots}
\title{Internal method replicating dplyr's compat_as_lazy_dots as it's not exported}
\usage{
compat_as_lazy_dots(...)
}
\arguments{
\item{...}{passed to rlang::quos}
}
\description{
Internal method replicating dplyr's compat_as_lazy_dots as it's not exported
}
