% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{count_lines_in_file}
\alias{count_lines_in_file}
\title{Count number of lines in file
First tries unix command wc.
If that fails, it tries line count functions from different packages.
Last resort is readLines, which could fail if the file size is large, 
because it requires reading it into memory. Chunks are used to attempt to avoid this.}
\usage{
count_lines_in_file(file)
}
\description{
Count number of lines in file
First tries unix command wc.
If that fails, it tries line count functions from different packages.
Last resort is readLines, which could fail if the file size is large, 
because it requires reading it into memory. Chunks are used to attempt to avoid this.
}
\seealso{
\code{\link[bigreadr]{nlines}}, 
         \code{\link[LaF]{determine_nlines}}, 
         \code{\link[R.utils]{countLines}}
}
\keyword{internal}
