% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{create_dplyr_mapper}
\alias{create_dplyr_mapper}
\title{A function to make it easier to create dplyr function for disk.frame}
\usage{
create_dplyr_mapper(dplyr_fn)
}
\description{
A function to make it easier to create dplyr function for disk.frame
}
