% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv.to.disk.frame.r
\name{csv.to.disk.frame}
\alias{csv.to.disk.frame}
\title{Convert a csv file to disk.frame format}
\usage{
csv.to.disk.frame(infile, outpath, ..., mem.prop = 0.1)
}
\arguments{
\item{infile}{The csv file to convert into disk.frame format}

\item{outpath}{The path (in the form of a folder) to output the disk.frame. A
disk.frame is folder containing multiple R data files}

\item{...}{Parameters passed to read.table}

\item{mem.prop}{The proportion of memory to use. On Windows the memory limit
is determined using the memory.limit() function while on other platforms
the avaialble memory is assumed to be 4Gb. So by default on Windows it will
use 10% of the memory while on Linux it will use 400mb}
}
\description{
Convert a csv file to disk.frame format
}

