% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{csv2fst}
\alias{csv2fst}
\title{Convert a csv file to an fst file}
\usage{
csv2fst(infile, nchunks, backend = c("data.table", "readr", "LaF"),
  chunk_reader = c("readr", "readLines", "LaF"), header_row = 1,
  inmapfn = base::I, outdir = tempfile(fileext = ".df"),
  compress = 50, .progress = TRUE, backend_args = backend_args)
}
\arguments{
\item{infile}{One or more csv files}

\item{nchunks}{Number of chunks to use when reading the csv file(s)}

\item{backend}{CSV file reader to use.}

\item{chunk_reader}{Method to use to read chunks.}

\item{inmapfn}{A function to be applied to the chunk read in from CSV before 
the chunk is being written out.}

\item{outdir}{Directory to save the fst files.}

\item{compress}{Compression level for fst files. See \code{\link[fst]{write_fst}}}

\item{.progress}{If TRUE, display progress when importing csv files.}

\item{backend_args}{Additional parameters for backend data.table, readr, or LaF.}
}
\description{
Convert a csv file to an fst file
}
\seealso{
\code{\link{csv_to_disk.frame}}
         \code{\link[readr]{read_delim_chunked}}
}
\keyword{internal}
