% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{csv2fst_data.table}
\alias{csv2fst_data.table}
\title{Convert csv file to fst file using data.table::fread.}
\usage{
csv2fst_data.table(csv_file, inmapfn = base::I, compress = 50,
  outdir = tempfile(fileext = ".df"), ...)
}
\description{
Convert csv file to fst file using data.table::fread.
}
\keyword{internal}
