% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{csv2fst_laf}
\alias{csv2fst_laf}
\title{convert csv to fst file using LaF::detect_dm_csv
re-save to file if a character vector is passed instead of a file name.}
\usage{
csv2fst_laf(csv_file, nchunks = 1, inmapfn = base::I,
  outdir = tempfile(fileext = ".df"), compress = 50, ...)
}
\description{
convert csv to fst file using LaF::detect_dm_csv
re-save to file if a character vector is passed instead of a file name.
}
\seealso{
\code{\link[LaF]{detect_dm_csv}}, 
         \code{\link[LaF]{laf_open}}, 
         \code{\link[LaF]{next_block}}
}
\keyword{internal}
