% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{csv2fst_readr}
\alias{csv2fst_readr}
\title{convert csv file to fst file using readr::read_delim}
\usage{
csv2fst_readr(csv_file, nchunks = 1, inmapfn = base::I,
  compress = 50, outdir = tempfile(fileext = ".df"), delim = ",",
  ...)
}
\description{
convert csv file to fst file using readr::read_delim
}
\seealso{
\code{\link[readr]{read_delim}}, 
         \code{\link[readr]{read_delim_chunked}}
}
\keyword{internal}
