% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{determine_classes_csv_data.table}
\alias{determine_classes_csv_data.table}
\title{Determine classes for one or more files using data.table::fread by reading in 
first row. In case of conflict between files, use the more general class.
(follows data.table approach)}
\usage{
determine_classes_csv_data.table(infiles, ...)
}
\description{
Determine classes for one or more files using data.table::fread by reading in 
first row. In case of conflict between files, use the more general class.
(follows data.table approach)
}
\keyword{internal}
