% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{determine_classes_csv_laf}
\alias{determine_classes_csv_laf}
\title{Read a file or files using LaF::detect_dm_csv and use the created model to 
get the column classes. In case of conflict between files, use the more general class.}
\usage{
determine_classes_csv_laf(infiles, ...)
}
\description{
Read a file or files using LaF::detect_dm_csv and use the created model to 
get the column classes. In case of conflict between files, use the more general class.
}
\keyword{internal}
