% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{determine_classes_csv_readr}
\alias{determine_classes_csv_readr}
\title{Read first row of a file or files using readr::read_delim and use the result to 
determine classes. In case of conflict between files, use the more general class.}
\usage{
determine_classes_csv_readr(infiles, delim = ",", ...)
}
\description{
Read first row of a file or files using readr::read_delim and use the result to 
determine classes. In case of conflict between files, use the more general class.
}
\keyword{internal}
