% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{df2fst}
\alias{df2fst}
\title{Saves a dataframe to an fst file.}
\usage{
df2fst(df, inmapfn = base::I, compress = 50, outdir = tempdir(),
  outfile_template = tempfile())
}
\description{
Saves a dataframe to an fst file.
}
\keyword{internal}
