% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{disk.frame}
\alias{disk.frame}
\title{Create a disk.frame}
\usage{
disk.frame(path, preview.rows = 1, ...)
}
\arguments{
\item{path}{The path to store the output file}

\item{preview.rows}{The number of rows to load into memory}
}
\description{
Create a disk.frame
}

