% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{disk.frame_folder}
\alias{disk.frame_folder}
\alias{disk.frame_fst}
\title{Create a disk.frame pointed to a folder}
\usage{
disk.frame_folder(path)

disk.frame_fst(path)
}
\arguments{
\item{path}{The path to store the output file or to a directory}
}
\description{
Create a disk.frame pointed to a folder
}
