% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame-to-parquet.r
\name{disk.frame_to_parquet}
\alias{disk.frame_to_parquet}
\title{A function to convert a disk.frame to parquet format}
\usage{
disk.frame_to_parquet(df, outdir)
}
\description{
A function to convert a disk.frame to parquet format
}
