% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{disk.lib}
\alias{disk.lib}
\title{Look inside the paths looking for disk.frame data sets. *** Stub ***}
\usage{
disk.lib(path)
}
\description{
Look inside the paths looking for disk.frame data sets. *** Stub ***
}

