% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{distribute}
\alias{distribute}
\title{Distribute (chunk-up/break-up) a fst file into chunks into a folder}
\usage{
distribute(...)
}
\arguments{
\item{df}{The disk.frame}

\item{outdirpath}{The directory/folder to output the chunks to}

\item{compress}{Level of compression: 0-100 where 100 is the highest level of compress}
}
\description{
Distribute (chunk-up/break-up) a fst file into chunks into a folder
}
