% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2disk.frame.r
\name{estimate_chunks_needed}
\alias{estimate_chunks_needed}
\title{Estimate number of chunks needed}
\usage{
estimate_chunks_needed(infile, max_percent_ram = 0.5, ...)
}
\arguments{
\item{infile}{One or more file locations, as character vector.}

\item{max_percent_ram}{Maximum percent RAM that should be devoted to the file import.}

\item{...}{Currently unused}
}
\value{
Integer vector, length of infile, indicating the number of chunks that should be used.
}
\description{
Estimate number of chunks needed for one or more files. Does this by assuming the file size is 
commensurate with amount of RAM required, and comparing to total RAM available.
Can be offset by setting max_percent_ram. In general, set max_percent_ram to well under 1.0 if
the files may be read in parallel or if you anticipate copies may be made.
}
