% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{evalparseglue}
\alias{evalparseglue}
\title{Helper function to evalparse some `glue::glue` string}
\usage{
evalparseglue(code, env = parent.frame())
}
\arguments{
\item{code}{the code in character(string) format to evaluate}

\item{env}{the environment in which to evaluate the code}
}
\description{
Helper function to evalparse some `glue::glue` string
}
