% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{find_globals_recursively}
\alias{find_globals_recursively}
\title{Find globals in an expression by searching through the chain}
\usage{
find_globals_recursively(code, envir)
}
\arguments{
\item{code}{An expression to search for globals}

\item{envir}{The environment from which to begin the search}
}
\description{
Find globals in an expression by searching through the chain
}
