% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_test_data.r
\name{gen_datatable_synthetic}
\alias{gen_datatable_synthetic}
\title{Generate synthetic dataset for testing}
\usage{
gen_datatable_synthetic(N = 2e+08, K = 100)
}
\arguments{
\item{N}{number of rows. Defaults to 200 million}

\item{K}{controls the number of unique values for id. Some ids will have K distinct values while others have N/K distinct values}
}
\description{
Generate synthetic dataset for testing
}
