% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-verbs.R
\name{generate_summ_code}
\alias{generate_summ_code}
\title{Helper function to generate summarisation code}
\usage{
generate_summ_code(...)
}
\description{
Helper function to generate summarisation code
}
