% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chunk_ids.r
\name{get_chunk_ids}
\alias{get_chunk_ids}
\title{Get the chunk file names}
\usage{
get_chunk_ids(df, ..., full.names = F, strip_extension = T)
}
\arguments{
\item{df}{a disk.frame}

\item{...}{passed to list.files}

\item{full.names}{If TRUE returns the full path to the file, Defaults to F.}

\item{strip_extension}{If TRUE then the file extension in the chunk_id is removed. Defaults to TRUE}
}
\description{
Get the chunk file names
}
