% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition.r
\name{get_partition_paths}
\alias{get_partition_paths}
\title{Get the partitioning structure of a folder}
\usage{
get_partition_paths(df)
}
\arguments{
\item{df}{a disk.frame whose paths will be used to determine if it's
folder-partitioned disk.frame}
}
\description{
Get the partitioning structure of a folder
}
