% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-verbs.R
\name{group_by_aggregation}
\alias{group_by_aggregation}
\title{A two-stage aggregation framework}
\usage{
group_by_aggregation(chunkfn, agg, finalize)
}
\arguments{
\item{chunkfn}{a function to apply to each chunk. The results are passed to a central location for aggregation}

\item{agg}{The aggregation function in the central location each chunk}

\item{finalize}{A function to run at the end}
}
\value{
a disk.frame.agg
}
\description{
A two-stage aggregation framework
}
\examples{

# defining the mean aggregation

chunkfn <- function(chunk) {
}
}
