% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{groups.disk.frame}
\alias{groups.disk.frame}
\title{The shard keys of the disk.frame}
\usage{
\method{groups}{disk.frame}(x)
}
\arguments{
\item{x}{a disk.frame}
}
\value{
character
}
\description{
The shard keys of the disk.frame
}
