% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{head.disk.frame}
\alias{head.disk.frame}
\alias{tail.disk.frame}
\title{Head of the disk.frame}
\usage{
\method{head}{disk.frame}(x, n = 6L, ...)

\method{tail}{disk.frame}(x, n = 6L, ...)
}
\arguments{
\item{x}{a disk.frame}

\item{n}{number of rows to include}

\item{...}{passed to base::head or base::tail}
}
\description{
Head of the disk.frame

tail of disk.frame
}
