% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{header_row_index}
\alias{header_row_index}
\title{Get header row}
\usage{
header_row_index(csv_file, method = c("data.table", "readr", "LaF"), ...)
}
\arguments{
\item{csv_file}{A single csv file to examine.}

\item{method}{Method to use to determine the column names. 
data.table uses \code{\link[data.table]{fread}}. 
readr uses \code{\link[readr]{read_delim}}.}

\item{...}{Additional arguments to pass to the method.}
}
\value{
Line number for the header if found; 0 otherwise.
}
\description{
Determine the line number for the header row in a file.
Uses a simplistic method of matching the column names to find the header row.
}
