% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{is_ready}
\alias{is_ready}
\title{The readiness of a disk.frame}
\usage{
is_ready(df)
}
\arguments{
\item{df}{a disk.frame}
}
\description{
Sometimes a disk.frame may be in use and this check whether it is ready from a long running non-blocking process.
}
