% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{line_first_match}
\alias{line_first_match}
\title{Look for a line that matches the pattern in the file.
Will first try using the unix grep command to search the file.
If that fails, it will fall back on reading in chunks of the file.}
\usage{
line_first_match(file, pattern, in_chunk_size = 1000)
}
\description{
Look for a line that matches the pattern in the file.
Will first try using the unix grep command to search the file.
If that fails, it will fall back on reading in chunks of the file.
}
\keyword{internal}
