% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{map}
\alias{map}
\alias{map_dfr}
\alias{map_dfr.default}
\alias{map_dfr.disk.frame}
\alias{map.disk.frame}
\alias{imap_dfr}
\alias{imap.disk.frame}
\alias{lazy}
\alias{lazy.disk.frame}
\alias{delayed}
\alias{delayed.disk.frame}
\alias{chunk_lapply}
\title{Apply the same function to all chunks}
\usage{
map(.x, .f, ...)

map_dfr(.x, .f, ..., .id = NULL)

\method{map_dfr}{default}(.x, .f, ..., .id = NULL)

\method{map_dfr}{disk.frame}(.x, .f, ..., .id = NULL, use.names = fill,
  fill = FALSE, idcol = NULL)

\method{map}{disk.frame}(.x, .f, ..., outdir = NULL, keep = NULL,
  chunks = nchunks(.x), compress = 50, lazy = TRUE,
  overwrite = FALSE, vars_and_pkgs = future::getGlobalsAndPackages(.f,
  envir = parent.frame()))

imap_dfr(.x, .f, ..., .id = NULL)

imap.disk.frame(.x, .f, outdir = NULL, keep = NULL,
  chunks = nchunks(.x), compress = 50, lazy = TRUE,
  overwrite = FALSE)

lazy(.x, .f, ...)

\method{lazy}{disk.frame}(.x, .f, ...)

delayed(.x, .f, ...)

\method{delayed}{disk.frame}(.x, .f, ...)

chunk_lapply(...)
}
\arguments{
\item{.x}{a disk.frame}

\item{.f}{a function to apply to each of the chunks}

\item{...}{for compatibility with `purrr::map`}

\item{.id}{not used}

\item{use.names}{for map_dfr's call to data.table::rbindlist. See data.table::rbindlist}

\item{fill}{for map_dfr's call to data.table::rbindlist. See data.table::rbindlist}

\item{idcol}{for map_dfr's call to data.table::rbindlist. See data.table::rbindlist}

\item{outdir}{the output directory}

\item{keep}{the columns to keep from the input}

\item{chunks}{The number of chunks to output}

\item{compress}{0-100 fst compression ratio}

\item{lazy}{if TRUE then do this lazily}

\item{overwrite}{if TRUE removes any existing chunks in the data}

\item{vars_and_pkgs}{variables and packages to send to a background session. This is typically automatically detected}
}
\description{
Apply the same function to all chunks

imap.disk.frame accepts a two argument function where the first argument is a disk.frame and the 
second is the chunk ID

`lazy` is convenience function to apply `.f` to every chunk

Lazy chunk_lapply wrapper
}
