% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.r
\name{map2}
\alias{map2}
\alias{map_by_chunk_id}
\title{Perform a function on both disk.frames .x and .y, each chunk of .x and .y gets run by .f(x.chunk, y.chunk)}
\usage{
map2(.x, .y, .f, ...)

map_by_chunk_id(x, y, fn, outdir)
}
\arguments{
\item{.x}{a disk.frame}

\item{.y}{a disk.frame}

\item{.f}{a function to be called on each chunk of x and y matched by chunk_id}

\item{...}{not used}

\item{outdir}{output directory}
}
\description{
Perform a function on both disk.frames .x and .y, each chunk of .x and .y gets run by .f(x.chunk, y.chunk)
}
