% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_by_chunk_id.r
\name{map_by_chunk_id}
\alias{map_by_chunk_id}
\title{Perform a function on both disk.frames x and y, each chunk of x and y gets run by fn(x.chunk, y.chunk)}
\usage{
map_by_chunk_id(x, y, fn, outdir)
}
\arguments{
\item{x}{a disk.frame}

\item{y}{a disk.frame}

\item{fn}{a function to be called on each chunk of x and y matched by chunk_id}

\item{outdir}{output directory}
}
\description{
Perform a function on both disk.frames x and y, each chunk of x and y gets run by fn(x.chunk, y.chunk)
}
