% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{nrow}
\alias{nrow}
\alias{nrow.default}
\alias{nrow.disk.frame}
\title{Number of rows of disk.frame}
\usage{
nrow(df, ...)

\method{nrow}{default}(df, ...)

\method{nrow}{disk.frame}(df, ...)
}
\arguments{
\item{df}{a disk.frame}

\item{...}{passed to base::nrow}
}
\description{
Number of rows of disk.frame
}
