% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-stage-verbs.R
\name{var_df.chunk_agg.disk.frame}
\alias{var_df.chunk_agg.disk.frame}
\alias{var_df.collected_agg.disk.frame}
\alias{sd_df.chunk_agg.disk.frame}
\alias{sd_df.collected_agg.disk.frame}
\alias{mean_df.chunk_agg.disk.frame}
\alias{mean_df.collected_agg.disk.frame}
\alias{sum_df.chunk_agg.disk.frame}
\alias{sum_df.collected_agg.disk.frame}
\alias{min_df.chunk_agg.disk.frame}
\alias{min_df.collected_agg.disk.frame}
\alias{max_df.chunk_agg.disk.frame}
\alias{max_df.collected_agg.disk.frame}
\alias{median_df.chunk_agg.disk.frame}
\alias{median_df.collected_agg.disk.frame}
\alias{n_df.chunk_agg.disk.frame}
\alias{n_df.collected_agg.disk.frame}
\alias{length_df.chunk_agg.disk.frame}
\alias{length_df.collected_agg.disk.frame}
\alias{any_df.chunk_agg.disk.frame}
\alias{any_df.collected_agg.disk.frame}
\alias{all_df.chunk_agg.disk.frame}
\alias{all_df.collected_agg.disk.frame}
\alias{n_distinct_df.chunk_agg.disk.frame}
\alias{n_distinct_df.collected_agg.disk.frame}
\alias{quantile_df.chunk_agg.disk.frame}
\alias{quantile_df.collected_agg.disk.frame}
\alias{IQR_df.chunk_agg.disk.frame}
\alias{IQR_df.collected_agg.disk.frame}
\title{One Stage function}
\usage{
var_df.chunk_agg.disk.frame(x, na.rm = FALSE)

var_df.collected_agg.disk.frame(listx)

sd_df.chunk_agg.disk.frame(x, na.rm = FALSE)

sd_df.collected_agg.disk.frame(listx)

mean_df.chunk_agg.disk.frame(x, na.rm = FALSE, ...)

mean_df.collected_agg.disk.frame(listx)

sum_df.chunk_agg.disk.frame(x, ...)

sum_df.collected_agg.disk.frame(listx, ...)

min_df.chunk_agg.disk.frame(x, ...)

min_df.collected_agg.disk.frame(listx, ...)

max_df.chunk_agg.disk.frame(x, ...)

max_df.collected_agg.disk.frame(listx, ...)

median_df.chunk_agg.disk.frame(x, ...)

median_df.collected_agg.disk.frame(listx, ...)

n_df.chunk_agg.disk.frame(...)

n_df.collected_agg.disk.frame(listx, ...)

length_df.chunk_agg.disk.frame(x, ...)

length_df.collected_agg.disk.frame(listx, ...)

any_df.chunk_agg.disk.frame(x, ...)

any_df.collected_agg.disk.frame(listx, ...)

all_df.chunk_agg.disk.frame(x, ...)

all_df.collected_agg.disk.frame(listx, ...)

n_distinct_df.chunk_agg.disk.frame(x, na.rm = FALSE, ...)

n_distinct_df.collected_agg.disk.frame(listx, ...)

quantile_df.chunk_agg.disk.frame(x, ...)

quantile_df.collected_agg.disk.frame(listx, ...)

IQR_df.chunk_agg.disk.frame(x, na.rm = FALSE, ...)

IQR_df.collected_agg.disk.frame(listx, ...)
}
\arguments{
\item{x}{the input}

\item{na.rm}{Remove NAs. TRUE of FALSE}

\item{listx}{a list}

\item{...}{additional options}
}
\description{
One Stage function

mean chunk_agg

mean collected_agg
}
