% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-filter.r
\name{partition_filter}
\alias{partition_filter}
\title{Filter the dataset based on folder partitions}
\usage{
partition_filter(x, ...)
}
\arguments{
\item{x}{a disk.frame}

\item{...}{filtering conditions for filtering the disk.frame at (folder) partition level}
}
\description{
Filter the dataset based on folder partitions
}
