% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{play}
\alias{play}
\title{Internal methods}
\usage{
play(.data, cmds = NULL)
}
\arguments{
\item{.data}{the disk.frame}

\item{cmds}{the list of function to play back}
}
\description{
Internal methods
}
