% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var_to_xgboost_scorecard.r
\name{print.xgdf_scorecard}
\alias{print.xgdf_scorecard}
\title{Print xgboost scorecard}
\usage{
\method{print}{xgdf_scorecard}(x, ...)
}
\arguments{
\item{x}{the Xgboost modelling result}

\item{...}{not used. Kept for compatibility}
}
\description{
Print xgboost scorecard
}
