% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_group_by.r
\name{progressbar}
\alias{progressbar}
\title{Show a progress bar of the action being performed}
\usage{
progressbar(df)
}
\arguments{
\item{df}{a disk.frame}
}
\description{
Show a progress bar of the action being performed
}
