% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.r
\name{pull.disk.frame}
\alias{pull.disk.frame}
\title{Pull a column from table similar to `dplyr::pull`.}
\usage{
\method{pull}{disk.frame}(.data, var = -1, name = NULL, ...)
}
\arguments{
\item{.data}{The disk.frame}

\item{var}{can be an positive or negative integer or a character/string. See dplyr::pull documentation}
}
\description{
Pull a column from table similar to `dplyr::pull`.
}
