% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{purrr_as_mapper}
\alias{purrr_as_mapper}
\title{Used to convert a function to purrr syntax if needed}
\usage{
purrr_as_mapper(.f)
}
\arguments{
\item{.f}{a normal function or purrr syntax function i.e. `~{ ...code...}`}
}
\description{
Used to convert a function to purrr syntax if needed
}
