% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindlist.disk.frame.r
\name{rbindlist.disk.frame}
\alias{rbindlist.disk.frame}
\title{rbindlist disk.frames together}
\usage{
rbindlist.disk.frame(df_list, outdir, by_chunk_id = T, parallel = T,
  compress = 50, overwrite = T)
}
\arguments{
\item{df_list}{A list of disk.frames}

\item{outdir}{Output directory of the row-bound disk.frames}

\item{by_chunk_id}{If TRUE then only the chunks with the same chunk IDs will be bound}

\item{parallel}{if TRUE then bind multiple disk.frame simultaneously, Defaults to TRUE}

\item{compress}{0-100, 100 being the highest compression rate.}

\item{overwrite}{overwrite the output directory}
}
\description{
rbindlist disk.frames together
}
