% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rechunk.r
\name{rechunk}
\alias{rechunk}
\title{Increase or decrease the number of chunks in the disk.frame}
\usage{
rechunk(df, nchunks, outdir = attr(df, "path"), shardby = NULL,
  overwrite = TRUE)
}
\arguments{
\item{df}{the disk.frame to rechunk}

\item{nchunks}{number of chunks}

\item{outdir}{the output directory}

\item{shardby}{the shardkeys}

\item{overwrite}{overwrite the output directory}
}
\description{
Increase or decrease the number of chunks in the disk.frame
}
\examples{
cars.df = as.disk.frame(cars, nchunks = 2)

# re-chunking cars.df to 3 chunks, done "in-place" to the same folder as cars.df
\dontrun{
rechunk(cars.df, 3)
}

new_path = tempfile(fileext = ".df")
# re-chunking cars.df to 4 chunks, shardby speed, done "out-of-place"
cars2.df = rechunk(cars.df, 4, new_path, shardby = "speed")

# clean up cars.df
delete(cars.df)
delete(cars2.df)
}
