% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{renumber_file_names}
\alias{renumber_file_names}
\title{number files in order from filename_suffix01, prefix_suffix02, etc.
numbering increased with number of files. 
so if 1000 files, number prefix_suffix0001, prefix_suffix0002, ..., prefix_suffix1000
Example:
renumber_file_names(LETTERS)
renumber_file_names(LETTERS, suffix = "split")
renumber_file_names(LETTERS, suffix = "split", extension = "csv")}
\usage{
renumber_file_names(files, suffix = "", extension = NULL)
}
\description{
number files in order from filename_suffix01, prefix_suffix02, etc.
numbering increased with number of files. 
so if 1000 files, number prefix_suffix0001, prefix_suffix0002, ..., prefix_suffix1000
Example:
renumber_file_names(LETTERS)
renumber_file_names(LETTERS, suffix = "split")
renumber_file_names(LETTERS, suffix = "split", extension = "csv")
}
\keyword{internal}
