% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_frac.R
\name{sample_frac.disk.frame}
\alias{sample_frac.disk.frame}
\title{Sample n rows from a disk.frame}
\usage{
\method{sample_frac}{disk.frame}(.data, ...)
}
\arguments{
\item{.data}{a disk.frame}

\item{...}{passed to dplyr::samle_frac}
}
\description{
Sample n rows from a disk.frame
}
\examples{
cars.df = as.disk.frame(cars)

collect(sample_frac(cars.df, 0.5))

# clean up cars.df
delete(cars.df)
}
