% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{setup_disk.frame}
\alias{setup_disk.frame}
\title{Set up disk.frame environment}
\usage{
setup_disk.frame(workers = parallel::detectCores(logical = FALSE),
  future_backend = multiprocess, future.globals.maxSize = Inf, ...,
  gui = FALSE)
}
\arguments{
\item{workers}{the number of workers (background R processes in the}

\item{future_backend}{which future backend to use for parallelization}

\item{future.globals.maxSize}{The amount of memory that is allowed to be
transferred between workers. Defaults to Inf for no limit.
multiprocess environment)}

\item{...}{passed to `future::plan`}

\item{gui}{Whether to use a Graphical User Interface (GUI) for selecting the options. Defaults to FALSE}
}
\description{
Set up disk.frame environment
}
\examples{
if (interactive()) {
  # use a Shiny GUI to adjust settings
  setup_disk.frame(gui = TRUE)

  # set the number workers to 2
  setup_disk.frame(2)

  # setup disk.frame to use multiple workers
  # these may use more than two cores, so it commented out
  # setup_disk.frame()

  # set the future backend to sequential
  # these may use more than two cores, so it commented out
  # setup_disk.frame(future_backend = future::sequential)
}
}
