% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shardkey.r
\name{shardkey_equal}
\alias{shardkey_equal}
\title{Compare two disk.frame shardkeys}
\usage{
shardkey_equal(sk1, sk2)
}
\arguments{
\item{sk1}{shardkey1}

\item{sk2}{shardkey2}
}
\description{
Compare two disk.frame shardkeys
}
