% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ceremony.R
\name{show_ceremony}
\alias{show_ceremony}
\alias{ceremony_text}
\alias{show_boilerplate}
\alias{insert_ceremony}
\title{Show the code to setup disk.frame}
\usage{
show_ceremony()

ceremony_text()

show_boilerplate()

insert_ceremony()
}
\description{
Show the code to setup disk.frame
}
