% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{split_csv_file}
\alias{split_csv_file}
\title{Split a file into chunks}
\usage{
split_csv_file(infile, num_splits = 2, header = TRUE,
  split_method = c("bigreadr", "unix"), outdir = tempdir(),
  suffix = "split")
}
\arguments{
\item{infile}{One or more csv files to split}

\item{num_splits}{Number of splits for each file. If more than one infile, can specify either
a single num_splits value to apply to every file, or a vector of integers with a length 
equal to the number of infiles.}

\item{header}{Either TRUE, FALSE, or an integer specifying the header index.
If TRUE, the first line of the file will be copied to each file split.
If an integer is provided, all lines above or equal to that integer 
will be copied to each file split.
 
If the header is not known, try using \code{\link{header_row_index}} to determine the correct header.}

\item{outdir}{Where to save the split files.}

\item{suffix}{Suffix to use when naming the split files.}

\item{method}{If data.table, uses \code{\link[data.table]{fread}} to guess at the header location.
If bigreadr, must specify either header TRUE or FALSE. 
\code{\link[bigreadr]{split_file}} used to split the file.
If header is not specified, it will be assumed to be TRUE.
If none, header must be specified.
Note that data.table relies on unix tools that may not be available on a Windows platform.}
}
\value{
Vector of file paths for the split files.
}
\description{
Split a csv file into multiple parts. 
Uses either bigreadr or data.table to split the file, adding the header to each split.
}
