% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{split_csv_file_bigreadr}
\alias{split_csv_file_bigreadr}
\title{Split csv file using \code{\link[bigreadr]{split_file}}
Requires that header is either TRUE or FALSE.
Does not handle blank spaces or other text before the header.}
\usage{
split_csv_file_bigreadr(csv_file, outdir = tempdir(), num_splits = 2,
  header = TRUE)
}
\description{
Split csv file using \code{\link[bigreadr]{split_file}}
Requires that header is either TRUE or FALSE.
Does not handle blank spaces or other text before the header.
}
\seealso{
\code{\link[bigreadr]{nlines}}, 
         \code{\link[bigreadr]{split_file}}, 
         \code{\link[bigreadr]{get_split_files}}
}
\keyword{internal}
