% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_csv_file.r
\name{split_csv_file_unix}
\alias{split_csv_file_unix}
\title{Split csv file using unix tools
Can split a file with any sized header.}
\usage{
split_csv_file_unix(csv_file, outdir = tempdir(), num_splits = 2,
  header_row_index = 1)
}
\description{
Split csv file using unix tools
Can split a file with any sized header.
}
\keyword{internal}
