% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition.r
\name{split_string_into_df}
\alias{split_string_into_df}
\title{Turn a string of the form /partion1=val/partion2=val2 into data.frame}
\usage{
split_string_into_df(path_strs)
}
\description{
Turn a string of the form /partion1=val/partion2=val2 into data.frame
}
