% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srckeep.disk.frame.r
\name{srckeep}
\alias{srckeep}
\title{Keep only the variables from the input listed in selections}
\usage{
srckeep(df, selections)
}
\arguments{
\item{df}{a disk.frame}

\item{selections}{The list of variables to keep from the input source}
}
\description{
Keep only the variables from the input listed in selections
}
