% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-verbs.R
\name{summarise.grouped_disk.frame}
\alias{summarise.grouped_disk.frame}
\title{#' @export
quantile.chunk_agg.disk.frame <- function(x, ...) {
  quantile(x, ...)
}}
\usage{
\method{summarise}{grouped_disk.frame}(.data, ...)
}
\description{
#' @export
quantile.collected_agg.disk.frame <- function(listx, ...) {
  quantile(unlist(listx), ...)
}
A function to parse the summarize function
}
