% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table.to.disk.frame.r
\name{table.to.disk.frame}
\alias{csv.to.disk.frame}
\alias{table.to.disk.frame}
\title{Convert a table file to disk.frame format}
\usage{
table.to.disk.frame(infile, outpath, ..., mem.prop = 0.1, VERBOSE = FALSE,
  read_fun = read.table)

csv.to.disk.frame(...)
}
\arguments{
\item{infile}{The csv file to convert into disk.frame format}

\item{outpath}{The path (in the form of a folder) to output the disk.frame. A
disk.frame is folder containing multiple R data files}

\item{...}{Parameters passed to read.fun and table.to.disk.frame}

\item{mem.prop}{The proportion of memory to use. On Windows the memory limit
is determined using the memory.limit() function while on other platforms
the avaialble memory is assumed to be 4Gb. So by default on Windows it will
use 10\% of the memory while on Linux it will use 400mb}

\item{VERBOSE}{Prints additional error messages}

\item{read_fun}{The function used to read in the table. The funciton needs t
have parameters nrows and header (as a boolean) arguments defined}
}
\description{
Convert a table file to disk.frame format
}

