% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfast-verbs.r
\name{chunk_dt_count.disk.frame}
\alias{chunk_dt_count.disk.frame}
\alias{chunk_dt_uncount.disk.frame}
\alias{chunk_dt_unnest}
\alias{chunk_dt_nest}
\alias{chunk_dt_hoist}
\alias{chunk_dt_fill}
\alias{dt_separate.disk.frame}
\title{The tidy verbs implemented for disk.frame}
\usage{
chunk_dt_count.disk.frame(.data, ...)

chunk_dt_uncount.disk.frame(.data, ...)

chunk_dt_unnest(.data, ...)

chunk_dt_nest(.data, ...)

chunk_dt_hoist(.data, ...)

chunk_dt_fill(.data, ...)

\method{dt_separate}{disk.frame}(.data, ...)
}
\arguments{
\item{.data}{a disk.frame}

\item{...}{Same as the tidyfast functions}
}
\description{
Please see the tidyfast document for their usage
}
\examples{
library(tidyfast)
library(data.table)

#' create a disk.frame
disk.frame_to_split <- as.disk.frame(data.table(
  x = paste(letters, LETTERS, sep = ".")
))

disk.frame_to_split \%>\% 
  dt_separate(x, into = c("lower", "upper")) \%>\% 
  collect

#' clean up
delete(disk.frame_to_split)
}
\concept{tidyfast verbs}
