% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.zip.to.sqlite.r
\name{zip.file.to.sqlite}
\alias{zip.file.to.sqlite}
\title{Convert a csv file to sqlite format}
\usage{
zip.file.to.sqlite(file, file.name, outpath, outtable, read_fun = read.csv,
  ..., safely = TRUE, VERBOSE = FALSE)
}
\arguments{
\item{file}{The zip file to convert into disk.frame format}

\item{file.name}{The file name of the file within the zip file to use}

\item{outpath}{The path (in the form of a folder) to output to sqlite. A
disk.frame is folder containing multiple R data files}

\item{outtable}{The output table name in sqlite}

\item{read_fun}{The function used to read the files. The default is read.csv}

\item{...}{Parameters passed to read_fun}

\item{safely}{Reads the file in chunks instead of in one go}

\item{VERBOSE}{Prints additionally messages or not}
}
\description{
Convert a csv file to sqlite format
}
