% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.r
\name{compute.disk.frame}
\alias{compute.disk.frame}
\title{Perform the computation}
\usage{
\method{compute}{disk.frame}(x, name, outdir, overwrite = T, ...)
}
\arguments{
\item{x}{a disk.frame}

\item{name}{not used}

\item{outdir}{the output directory}

\item{overwrite}{if TRUE overwrite outdir if it's a disk.frame}

\item{...}{not used}
}
\description{
Perform the computation
}
