% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{create_dplyr_mapper}
\alias{create_dplyr_mapper}
\title{Create dplyr function for disk.frame}
\usage{
create_dplyr_mapper(dplyr_fn, warning_msg = NULL)
}
\arguments{
\item{dplyr_fn}{The dplyr function to create a mapper for}

\item{warning_msg}{The warning message to display when invoking the mapper}
}
\description{
A function to make it easier to create functions like \code{filter}
}
\examples{


}
