% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{create_dplyr_mapper}
\alias{create_dplyr_mapper}
\title{Kept for backwards-compatibility to be removed in 0.3}
\usage{
create_dplyr_mapper()
}
\description{
Kept for backwards-compatibility to be removed in 0.3
}
