% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{disk.frame}
\alias{disk.frame}
\title{Create a disk.frame}
\usage{
disk.frame(path, ..., backend = "fst")
}
\arguments{
\item{path}{The path to store the output file or to a directory}

\item{preview.rows}{The number of rows to load into memory}
}
\description{
Create a disk.frame
}
