% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{disk.frame}
\alias{disk.frame}
\title{Create a disk.frame}
\usage{
disk.frame(path, backend = "fst")
}
\arguments{
\item{path}{The path to store the output file or to a directory}

\item{backend}{The only available backend is fst at the moment}
}
\description{
Create a disk.frame
}
\examples{
path = outdir=file.path(tempdir(),"cars")
as.disk.frame(cars, path, overwrite = T, nchunks = 2)
df = disk.frame(path)
head(df)
nchunks(df)
}
