% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame-to-parquet.r
\name{disk.frame_to_parquet}
\alias{disk.frame_to_parquet}
\title{A function to convert a disk.frame to parquet format}
\usage{
disk.frame_to_parquet(df, outdir)
}
\arguments{
\item{df}{a disk.frame or a path to a disk.frame}

\item{outdir}{the path to save the parquet files}
}
\description{
A function to convert a disk.frame to parquet format
}
