% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{select.disk.frame}
\alias{select.disk.frame}
\alias{rename_.disk.frame}
\alias{filter_.disk.frame}
\alias{mutate_.disk.frame}
\alias{transmute_.disk.frame}
\alias{arrange_.disk.frame}
\alias{summarise_.disk.frame}
\alias{do_.disk.frame}
\alias{glimpse.disk.frame}
\title{dplyr verbs implemented for disk.frame}
\usage{
\method{select}{disk.frame}(.data, ...)

\method{rename_}{disk.frame}(.data, ..., .dots)

\method{filter_}{disk.frame}(.data, ..., .dots)

\method{mutate_}{disk.frame}(.data, ..., .dots)

\method{transmute_}{disk.frame}(.data, ..., .dots)

\method{arrange_}{disk.frame}(.data, ..., .dots)

\method{summarise_}{disk.frame}(.data, ..., .dots)

\method{do_}{disk.frame}(.data, ..., .dots)

\method{glimpse}{disk.frame}(.data, ...)
}
\arguments{
\item{.data}{a disk.frame}

\item{...}{Same as the dplyr functions}
}
\description{
dplyr verbs implemented for disk.frame

mutate

transmuate

Take a glimpse
}
