% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{find_globals_recursively}
\alias{find_globals_recursively}
\title{Find globals in an expression by searching through the chain}
\usage{
find_globals_recursively(code, envir)
}
\description{
Find globals in an expression by searching through the chain
}
