% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.r
\name{group_by.disk.frame}
\alias{group_by.disk.frame}
\title{Group by designed for disk.frames}
\usage{
\method{group_by}{disk.frame}(.data, ...)
}
\arguments{
\item{.data}{a disk.frame}

\item{...}{same as the dplyr::group_by}
}
\description{
Group by designed for disk.frames
}
