% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-stage-verbs.R
\name{summarise.grouped_disk.frame}
\alias{summarise.grouped_disk.frame}
\alias{summarize.grouped_disk.frame}
\alias{group_by.disk.frame}
\alias{summarize.disk.frame}
\alias{summarise.disk.frame}
\title{A function to parse the summarize function}
\usage{
\method{summarise}{grouped_disk.frame}(.data, ...)

\method{summarize}{grouped_disk.frame}(.data, ...)

\method{group_by}{disk.frame}(.data, ..., add = FALSE, .drop = dplyr::group_by_drop_default(.data))

\method{summarize}{disk.frame}(.data, ...)

\method{summarise}{disk.frame}(.data, ...)
}
\arguments{
\item{.data}{a disk.frame}

\item{...}{same as the dplyr::group_by}

\item{add}{from dplyr}

\item{.drop}{from dplyr}
}
\description{
The disk.frame group by operation perform group WITHIN each chunk. This is
often used for performance reasons. If the user wishes to perform group-by,
they may choose to use the `hard_group_by` function which is expensive as it
reorganizes the chunks by the shard key.
}
\seealso{
hard_group_by
}
