% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_group_by.r
\name{hard_group_by}
\alias{hard_group_by}
\alias{hard_group_by.disk.frame}
\title{Perform a group by and ensuring that every unique grouping of by is
in the same chunk}
\usage{
hard_group_by(df, by, ...)

\method{hard_group_by}{disk.frame}(df, by,
  outdir = tempfile("tmp_disk_frame_hard_group_by"),
  nchunks = disk.frame::nchunks(df), overwrite = T, ...)
}
\arguments{
\item{df}{a disk.frame}

\item{by}{the columns to shard by}

\item{...}{not used}

\item{outdir}{the output directory}

\item{nchunks}{The number of chunks in the output. Defaults = nchunks.disk.frame(df)}

\item{overwrite}{overwrite the out put directory}
}
\description{
Perform a group by and ensuring that every unique grouping of by is
in the same chunk
}
