% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{is.file.disk.frame}
\alias{is.file.disk.frame}
\alias{is.dir.disk.frame}
\title{Checks if the df is a single-file based disk.frame}
\usage{
is.file.disk.frame(df, check.consistency = T)

is.dir.disk.frame(df, check.consistency = T)
}
\arguments{
\item{df}{a disk.frame}

\item{check.consistency}{check for consistency e.g. if it's actually a file}
}
\description{
Checks if the df is a single-file based disk.frame
}
