% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_disk.frame.r
\name{is_disk.frame}
\alias{is_disk.frame}
\title{Checks if a folder is a disk.frame}
\usage{
is_disk.frame(df)
}
\arguments{
\item{df}{a disk.frame or directory to check}
}
\description{
Checks if a folder is a disk.frame
}
