% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{is_ready}
\alias{is_ready}
\title{is the disk.frame ready from a long running non-blocking process}
\usage{
is_ready(df)
}
\arguments{
\item{df}{a disk.frame}
}
\description{
is the disk.frame ready from a long running non-blocking process
}
