% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{map}
\alias{map}
\alias{map.disk.frame}
\alias{imap.disk.frame}
\alias{lazy}
\alias{lazy.disk.frame}
\alias{delayed}
\alias{delayed.disk.frame}
\alias{chunk_lapply}
\title{Apply the same function to all chunks}
\usage{
map(.x, .f, ...)

\method{map}{disk.frame}(.x, .f, ..., outdir = NULL, keep = NULL,
  chunks = nchunks(.x), compress = 50, lazy = T, overwrite = F)

imap.disk.frame(.x, .f, outdir = NULL, keep = NULL,
  chunks = nchunks(.x), compress = 50, lazy = T, overwrite = F)

lazy(.x, .f, ...)

\method{lazy}{disk.frame}(.x, .f, ...)

delayed(.x, .f, ...)

\method{delayed}{disk.frame}(.x, .f, ...)

chunk_lapply(...)
}
\arguments{
\item{.x}{a disk.frame}

\item{.f}{a function to apply to each of the chunks}

\item{...}{for compatibility with `purrr::map`}

\item{outdir}{the output directory}

\item{keep}{the columns to keep from the input}

\item{chunks}{The number of chunks to output}

\item{compress}{0-100 fst compression ratio}

\item{lazy}{if TRUE then do this lazily}

\item{overwrite}{if TRUE removes any existing chunks in the data}
}
\description{
Apply the same function to all chunks

imap.disk.frame accepts a two argument function where the first argument is a disk.frame and the 
second is the chunk ID

`lazy` is convenience function to apply `.f` to every chunk

Lazy chunk_lapply wrapper
}
