% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchunks.r
\name{nchunks}
\alias{nchunks}
\alias{nchunk}
\alias{nchunk.disk.frame}
\alias{nchunks.disk.frame}
\title{Return the number of chunks}
\usage{
nchunks(df, ...)

nchunk(df, ...)

\method{nchunk}{disk.frame}(df, ...)

\method{nchunks}{disk.frame}(df, skip.ready.check = FALSE, ...)
}
\arguments{
\item{df}{a disk.frame}

\item{...}{not used}

\item{skip.ready.check}{NOT implemented}
}
\description{
Return the number of chunks

Returns the number of chunks in a disk.frame
}
