% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol-nrow.r
\name{nrow}
\alias{nrow}
\alias{nrow.disk.frame}
\alias{ncol}
\alias{ncol.disk.frame}
\title{Number of rows or columns}
\usage{
nrow(df, ...)

\method{nrow}{disk.frame}(df, ...)

ncol(df)

\method{ncol}{disk.frame}(df)
}
\arguments{
\item{df}{a disk.frame}

\item{...}{passed to base::nrow}
}
\description{
Number of rows or columns
}
\examples{
cars.df = as.disk.frame(cars)

# return total number of column and rows
ncol(cars.df)
nrow(cars.df)

# clean up cars.df
delete(cars.df)
}
