% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{nrow}
\alias{nrow}
\alias{nrow.default}
\alias{nrow.disk.frame}
\title{Number of rows of disk.frame}
\usage{
nrow(...)

\method{nrow}{default}(...)

\method{nrow}{disk.frame}(df)
}
\description{
Number of rows of disk.frame
}
