% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_check.r
\name{overwrite_check}
\alias{overwrite_check}
\title{Check if the outdir exists or not; and cleans it if it's not already there}
\usage{
overwrite_check(outdir, overwrite)
}
\arguments{
\item{outdir}{the output directory}

\item{overwrite}{TRUE or FALSE if `outdir`` exists and overwrite = FALSE then throw an error}
}
\description{
Check if the outdir exists or not; and cleans it if it's not already there
}
\examples{
tf = tempfile()
overwrite_check(tf, overwrite = FALSE)
overwrite_check(tf, overwrite = TRUE)

# clean up
fs::dir_delete(tf)
}
