% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.r
\name{play}
\alias{play}
\title{Play the recorded lazy operations}
\usage{
play(dataframe, recordings)
}
\arguments{
\item{dataframe}{A data.frame}

\item{recordings}{A recording the expression, globals and packages using create_chunk_mapper}
}
\description{
Play the recorded lazy operations
}
