% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.disk.frame.r
\name{print.disk.frame}
\alias{print.disk.frame}
\title{a new print method for disk.frame}
\usage{
\method{print}{disk.frame}(x, ...)
}
\arguments{
\item{x}{disk.frame}

\item{...}{not used}
}
\description{
a new print method for disk.frame
}
