% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rechunk.r
\name{rechunk}
\alias{rechunk}
\title{Increase or decrease the number of chunks in the disk.frame}
\usage{
rechunk(df, nchunks, outdir = attr(df, "path"), shardby = NULL,
  overwrite = T)
}
\arguments{
\item{df}{the disk.frame to rechunk}

\item{nchunks}{number of chunks}

\item{outdir}{the output directory}

\item{shardby}{the shardkeys}

\item{overwrite}{overwrite the output directory}
}
\description{
Increase or decrease the number of chunks in the disk.frame
}
