% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{setup_disk.frame}
\alias{setup_disk.frame}
\title{Set up disk.frame environment}
\usage{
setup_disk.frame(future_backend = multiprocess,
  workers = parallel::detectCores(logical = F), ...)
}
\arguments{
\item{future_backend}{which future backend to use for parallelisation}

\item{workers}{the number of workers (background R processes in the multiprocess environment)}

\item{...}{passed to `future::plan`}
}
\description{
Set up disk.frame environment
}
