% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table.r
\name{[.disk.frame}
\alias{[.disk.frame}
\title{[ interface for disk.frame using fst backend}
\usage{
\method{[}{disk.frame}(df, i, j, ..., keep = NULL)
}
\arguments{
\item{df}{a disk.frame}

\item{i}{same as data.table}

\item{j}{same as data.table}

\item{...}{same as data.table}

\item{keep}{the columns to srckeep}
}
\description{
[ interface for disk.frame using fst backend
}
