% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.frame.r
\name{[.disk.frame}
\alias{[.disk.frame}
\title{[ interface for disk.frame using fst backend}
\usage{
\method{[}{disk.frame}(df, i, j, ..., keep = NULL)
}
\description{
[ interface for disk.frame using fst backend
}
