% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table.r
\name{[.disk.frame}
\alias{[.disk.frame}
\title{[ interface for disk.frame using fst backend}
\usage{
\method{[}{disk.frame}(
  df,
  ...,
  keep = NULL,
  rbind = TRUE,
  use.names = TRUE,
  fill = FALSE,
  idcol = NULL
)
}
\arguments{
\item{df}{a disk.frame}

\item{...}{same as data.table}

\item{keep}{the columns to srckeep}

\item{rbind}{Whether to rbind the chunks. Defaults to TRUE}

\item{use.names}{Same as in data.table::rbindlist}

\item{fill}{Same as in data.table::rbindlist}

\item{idcol}{Same as in data.table::rbindlist}
}
\description{
[ interface for disk.frame using fst backend
}
\examples{
cars.df = as.disk.frame(cars)
speed_limit = 50
cars.df[speed < speed_limit ,.N, cut(dist, pretty(dist))]

# clean up
delete(cars.df)
}
