% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-verbs.R
\name{summarise.grouped_disk.frame}
\alias{summarise.grouped_disk.frame}
\title{A function to parse the summarize function}
\usage{
\method{summarise}{grouped_disk.frame}(.data, ...)
}
\description{
A function to parse the summarize function
}
