% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_vars.r
\name{tbl_vars.disk.frame}
\alias{tbl_vars.disk.frame}
\alias{group_vars.disk.frame}
\title{Column names for RStudio auto-complete}
\usage{
\method{tbl_vars}{disk.frame}(x)

\method{group_vars}{disk.frame}(x)
}
\arguments{
\item{x}{a disk.frame}
}
\description{
Returns the names of the columns. Needed for RStudio to complete variable
names
}
