% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_disk.frame.r
\name{write_disk.frame}
\alias{write_disk.frame}
\alias{output_disk.frame}
\title{Write out a data.frame/disk.frame to a disk.frame location}
\usage{
write_disk.frame(df, outdir, nchunks, overwrite, shardkey, shardchunks,
  compress = 50, ...)

output_disk.frame(...)
}
\arguments{
\item{df}{a disk.frame}

\item{outdir}{output directory for the disk.frame}

\item{nchunks}{number of chunks}

\item{overwrite}{overwrite output directory}

\item{shardkey}{the columns to shard by}

\item{shardchunks}{the number of chunks to shard to shard to. It is always greater than or equal to nchunks}

\item{compress}{compression ratio for fst files}

\item{...}{passed to map.disk.frame}
}
\description{
Write out a data.frame/disk.frame to a disk.frame location
}
