% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_disk.frame.r
\name{write_disk.frame}
\alias{write_disk.frame}
\alias{output_disk.frame}
\title{Write out a data.frame/disk.frame to a disk.frame location. If df is a
data.frame, then df must contain the column .out.disk.frame.id. This is
intended to be a low-level version of writing disk.frames. Using the
as.disk.frame function is recommended for most cases}
\usage{
write_disk.frame(df, outdir, nchunks, overwrite, shardby = NULL,
  compress = 50, ...)

output_disk.frame(...)
}
\arguments{
\item{df}{a disk.frame}

\item{outdir}{output directory for the disk.frame}

\item{nchunks}{number of chunks}

\item{overwrite}{overwrite output directory}

\item{shardby}{the columns to shard by}

\item{compress}{compression ratio for fst files}

\item{...}{passed to map.disk.frame}
}
\description{
Write out a data.frame/disk.frame to a disk.frame location. If df is a
data.frame, then df must contain the column .out.disk.frame.id. This is
intended to be a low-level version of writing disk.frames. Using the
as.disk.frame function is recommended for most cases
}
